Update upstream source

+++ bsd-setres_id.c
@@ -38,20 +38,20 @@ setresgid(gid_t rgid, gid_t egid, gid_t sgid)
 #if defined(HAVE_SETREGID) && !defined(BROKEN_SETREGID)
 	if (setregid(rgid, egid) < 0) {
 		saved_errno = errno;
-		warnx("setregid %u: %.100s", rgid, strerror(errno));
+		warnx("setregid %lu: %.100s", (u_long)rgid, strerror(errno));
 		errno = saved_errno;
 		ret = -1;
 	}
 #else
 	if (setegid(egid) < 0) {
 		saved_errno = errno;
-		warnx("setegid %u: %.100s", (u_int)egid, strerror(errno));
+		warnx("setegid %lu: %.100s", (u_long)egid, strerror(errno));
 		errno = saved_errno;
 		ret = -1;
 	}
 	if (setgid(rgid) < 0) {
 		saved_errno = errno;
-		warnx("setgid %u: %.100s", rgid, strerror(errno));
+		warnx("setgid %lu: %.100s", (u_long)rgid, strerror(errno));
 		errno = saved_errno;
 		ret = -1;
 	}
@@ -73,7 +73,7 @@ setresuid(uid_t ruid, uid_t euid, uid_t suid)
 #if defined(HAVE_SETREUID) && !defined(BROKEN_SETREUID)
 	if (setreuid(ruid, euid) < 0) {
 		saved_errno = errno;
-		warnx("setreuid %u: %.100s", ruid, strerror(errno));
+		warnx("setreuid %lu: %.100s", (u_long)ruid, strerror(errno));
 		errno = saved_errno;
 		ret = -1;
 	}
@@ -82,14 +82,14 @@ setresuid(uid_t ruid, uid_t euid, uid_t suid)
 # ifndef SETEUID_BREAKS_SETUID
 	if (seteuid(euid) < 0) {
 		saved_errno = errno;
-		warnx("seteuid %u: %.100s", euid, strerror(errno));
+		warnx("seteuid %lu: %.100s", (u_long)euid, strerror(errno));
 		errno = saved_errno;
 		ret = -1;
 	}
 # endif
 	if (setuid(ruid) < 0) {
 		saved_errno = errno;
-		warnx("setuid %u: %.100s", ruid, strerror(errno));
+		warnx("setuid %lu: %.100s", (u_long)ruid, strerror(errno));
 		errno = saved_errno;
 		ret = -1;
 	}
